WT-2138 Skip temporary/pre-alloc log files when copying.

+++ test/suite/test_txn10.py
@@ -56,7 +56,10 @@ class test_txn10(wttest.WiredTigerTestCase, suite_subprocess):
         for fname in os.listdir(olddir):
             fullname = os.path.join(olddir, fname)
             # Skip lock file on Windows since it is locked
-            if os.path.isfile(fullname) and "WiredTiger.lock" not in fullname:
+            if os.path.isfile(fullname) and \
+                "WiredTiger.lock" not in fullname and \
+                "Tmplog" not in fullname and \
+                "Preplog" not in fullname:
                 shutil.copy(fullname, newdir)
         # close the original connection and open to new directory
         self.close_conn()
